Adds a "continue-on-error" parameter to the file transfer jobs
- when files or directories are not accessible (e.g. dynamic content or IO problems)
  we need more robustness. this means, the handling of an directory should not abort
  in case of an error on one file in the directory
- by default setting, nothing changes
- this only applies when the file transfer is running in a job, for interactive call from users
  in the backend, we neither want to offer this parameter, nor do we want to collect errors
  and show them in the regular user messages view.

Fixes: SE-13602

+++ src/main/java/sirius/biz/storage/layer3/Transfer.java
@@ -13,6 +13,7 @@ import sirius.biz.process.ProcessContext;
 import sirius.biz.storage.layer2.Blob;
 import sirius.biz.storage.util.Attempt;
 import sirius.biz.storage.util.StorageUtils;
+import sirius.kernel.commons.Monoflop;
 import sirius.kernel.commons.Value;
 import sirius.kernel.di.std.Part;
 import sirius.kernel.health.Exceptions;
@@ -45,6 +46,7 @@ public class Transfer {
     private boolean forceWrite = true;
     private ProcessContext processContext;
     private boolean autobatch = false;
+    private boolean continueOnError = false;
 
     @Part
     private static Jobs jobs;
@@ -124,6 +126,21 @@ public class Transfer {
     @CheckReturnValue
     public Transfer autobatch() {
         this.autobatch = true;
+        return this;
+    }
+
+    /**
+     * Marks the transfer to continue on errors.
+     * <p>
+     * This setting only applies when the transfer is executed with an {@linkplain ProcessContext process context} set.
+     * Then, the failures can be collected and then displayed properly.
+     * For interactive calls, when a failure occurs, the exception will be propagated and this setting gets ignored.
+     *
+     * @return the transfer helper itself for fluent method calls
+     */
+    @CheckReturnValue
+    public Transfer continueOnError() {
+        this.continueOnError = true;
         return this;
     }
 
@@ -345,18 +362,32 @@ public class Transfer {
             destinationDirectory.createAsDirectory();
         }
 
+        Monoflop errorWasIgnored = Monoflop.create();
+
         sourceDirectory.allChildren().stream().forEach(child -> {
-            if (child.isFile()) {
-                transferFileTo(child, destinationDirectory.findChild(child.name()), deleteContent);
-            } else {
-                // Transfer and delete the child directory in case content should be deleted.
-                transferDirectory(child, destinationDirectory.findChild(child.name()), deleteContent, deleteContent);
-            }
-            if (deleteContent) {
-                child.delete();
+            try {
+                if (child.isFile()) {
+                    transferFileTo(child, destinationDirectory.findChild(child.name()), deleteContent);
+                } else {
+                    // Transfer and delete the child directory in case content should be deleted.
+                    transferDirectory(child,
+                                      destinationDirectory.findChild(child.name()),
+                                      deleteContent,
+                                      deleteContent);
+                }
+                if (deleteContent) {
+                    child.delete();
+                }
+            } catch (Exception exception) {
+                if (continueOnError && processContext != null) {
+                    processContext.handle(exception);
+                    errorWasIgnored.toggle();
+                } else {
+                    throw exception;
+                }
             }
         });
-        if (deleteSourceDirectory) {
+        if (deleteSourceDirectory && !errorWasIgnored.isToggled()) {
             sourceDirectory.delete();
         }
     }
@@ -365,7 +396,6 @@ public class Transfer {
         if (!forceTransfer && !shouldTransfer(sourceFile, destinationFile)) {
             return;
         }
-
         for (Attempt attempt : Attempt.values()) {
             try (InputStream input = sourceFile.createInputStream()) {
                 destinationFile.consumeStream(input, sourceFile.size());
